Refactor AddFeedback component for performance
- Created separate components for the different form fields in order to
  reduce he amount of code in the AddFeedback component
- Moved dropdown state and activeOptionText state closer to their
  respective components
