+++ src/pages/AddFeedbackPage/components/AddFeedback/AddFeedback.jsx
@@ -1,16 +1,13 @@
-import { useState, useEffect, useRef } from 'react';
+import { useState } from 'react';
 import { useNavigate } from 'react-router-dom';
 import { GoBackBtn } from '../../../../components/GoBackBtn';
-import { Dropdown } from './SelectOption';
-import { InputDescription } from '../InputDescription';
+import { FeatureType } from './FeatureType';
+import { TitleInput } from './TitleInput';
+import { DetailInput } from './DetailInput';
 import plusSign from '../../../../assets/icons/new-feedback.svg';
-import arrowUp from '../../../../assets/icons/arrow-up.svg';
-import arrowDown from '../../../../assets/icons/arrow-down.svg';
 import styles from './_addFeedback.module.scss';
 
 export const AddFeedback = () => {
-  const [activeOptionText, setActiveOptionText] = useState('bug');
-  const [isDropDownOpen, setIsDropDownOpen] = useState(false);
   const [formData, setFormData] = useState({
     title: '',
     category: '',
@@ -18,8 +15,6 @@ export const AddFeedback = () => {
   });
 
   const navigate = useNavigate();
-  const containerRef = useRef(null);
-  const arrow = isDropDownOpen ? arrowUp : arrowDown;
 
   const handleInputData = (e) => {
     setFormData((prevData) => ({
@@ -32,45 +27,6 @@ export const AddFeedback = () => {
     navigate('/', { replace: true });
     return null;
   };
-
-  const handleKeyPress = (e) => {
-    if (e.key === 'Enter') {
-      setFormData((prevData) => ({
-        ...prevData,
-        category: e.target.innerText,
-      }));
-      setIsDropDownOpen(!isDropDownOpen);
-    }
-  };
-
-  const handleDropDown = () => setIsDropDownOpen(!isDropDownOpen);
-
-  const formControls = {
-    isDropDownOpen,
-    setIsDropDownOpen,
-    activeOptionText,
-    setActiveOptionText,
-    setFormData,
-    handleKeyPress,
-  };
-
-  useEffect(() => {
-    const clickOutsideDropdown = (e) => {
-      if (
-        isDropDownOpen &&
-        containerRef.current &&
-        !containerRef.current.contains(e.target)
-      ) {
-        setIsDropDownOpen(false);
-      }
-    };
-
-    document.addEventListener('mousedown', clickOutsideDropdown);
-
-    return () => {
-      document.removeEventListener('mousedown', clickOutsideDropdown);
-    };
-  }, [isDropDownOpen]);
 
   return (
     <div className={styles.container}>
@@ -91,67 +47,15 @@ export const AddFeedback = () => {
               onSubmit={handleSubmit}
               className={styles.form}
             >
-              <div className={styles.inputBox}>
-                <InputDescription
-                  label="feedback title"
-                  description="Add a short, descriptive headline"
-                  htmlFor="feedbackTitle"
-                  styles={styles}
-                />
-                <input
-                  id="feedbackTitle"
-                  className={styles.inputText}
-                  type="text"
-                  onChange={handleInputData}
-                  name="title"
-                  value={formData.title}
-                />
-              </div>
-
-              <div className={styles.inputBox}>
-                <InputDescription
-                  label="Category"
-                  description="Choose a category for your feedback"
-                  htmlFor="featureDropdown"
-                  styles={styles}
-                />
-
-                <div
-                  id="featureDropdown"
-                  className={styles.dropDownContainer}
-                  tabIndex="0"
-                  onKeyPress={handleKeyPress}
-                  ref={containerRef}
-                >
-                  <div
-                    className={styles.activeOptionContainer}
-                    onClick={handleDropDown}
-                  >
-                    <span className={styles.activeOptionText}>
-                      {activeOptionText}
-                    </span>
-                    <img className={styles.arrow} src={arrow} alt="" />
-                  </div>
-
-                  {isDropDownOpen && <Dropdown formControls={formControls} />}
-                </div>
-              </div>
-
-              <div className={styles.inputBox}>
-                <InputDescription
-                  label="feedback detail"
-                  description="Include any specific comments on what should be improved, added, etc."
-                  htmlFor="detailTextArea"
-                  styles={styles}
-                />
-                <textarea
-                  id="detailTextArea"
-                  className={styles.textArea}
-                  onChange={handleInputData}
-                  value={formData.detail}
-                  name="detail"
-                ></textarea>
-              </div>
+              <TitleInput
+                handleInputData={handleInputData}
+                formData={formData}
+              />
+              <FeatureType setFormData={setFormData} />
+              <DetailInput
+                handleInputData={handleInputData}
+                formData={formData}
+              />
             </form>
             <div className={styles.buttonBox}>
               <button
